refactor: move listener code to listener.go
